Signed-off-by: CachorroC <juankpato87@gmail.com>
a Please enter the commit message for your changes. Lines starting
a with 'a' will be ignored, and an empty message aborts the commit.
a
a On branch main
a Your branch is behind 'origin/main' by 1 commit, and can be fast-forwarded.
a   (use "git pull" to update your local branch)
a
a Changes to be committed:
a	modified:   src/models/idk.ts
a